+++ rinad/src/ipcp/flow-allocator.h
@@ -384,7 +384,7 @@ private:
 
 class TearDownFlowTimerTask: public rina::TimerTask {
 public:
-	static const double DELAY;
+	static const long DELAY;
 
 	TearDownFlowTimerTask(FlowAllocatorInstance * flow_allocator_instance,
 			const std::string& flow_object_name, bool requestor);
